+++ src/cells/ProductCell.h
@@ -44,10 +44,6 @@ public:
   std::unique_ptr<Cell> Copy(GroupCell *group) const override;
   const CellTypeInfo &GetInfo() override;
 
-  using Cell::SetCurrentPoint;
-  void SetCurrentPoint(wxPoint point) const override;
-  void Draw(wxDC *dc, wxDC *antialiassingDC) override;
-
 protected:
   //! What maxima command name corresponds to this cell?
   virtual const wxString GetMaximaCommandName() const override;